[Update] Remove Build warnings
- Explicit conversions where implicit ones where generating warnings
- Return a copy and not a reference

+++ ECell_Engine/headers/ReversibleRNG.hpp
@@ -24,7 +24,7 @@ namespace ECellEngine::Maths
 		*/
 		inline const float Exponential(const float& _lambda)
 		{
-			return -log(random_uniform(&rng)) / _lambda;
+			return (const float) - log(random_uniform(&rng)) / _lambda;
 		}
 
 		/*
